This needed a couple of extra fixes to qfcc to deal with casting between
integer types and handles.

+++ libs/util/mdfour.c
@@ -156,7 +156,7 @@ copy64 (uint32_t * M, const unsigned char *in)
 	int         i;
 
 	for (i = 0; i < 16; i++)
-		M[i] = (in[i * 4 + 3] << 24) | (in[i * 4 + 2] << 16) |
+		M[i] = ((uint32_t) in[i * 4 + 3] << 24) | (in[i * 4 + 2] << 16) |
 			(in[i * 4 + 1] << 8) | (in[i * 4 + 0] << 0);
 }
